chore: move data fetching to top level of app
