+++ src/views/List.jsx
@@ -1,24 +1,6 @@
-import { useEffect, useState } from 'react';
-import { streamListItems } from '../api';
 import { ListItem } from '../components/ListItem';
 
-export function List({ listToken }) {
-	const [data, setData] = useState([]);
-
-	useEffect(() => {
-		if (!listToken) return;
-
-		return streamListItems(listToken, (querySnap) => {
-			const nextData = [];
-			querySnap.docs.forEach((docRef) => {
-				const data = docRef.data();
-				data.id = docRef.id;
-				nextData.push(data);
-			});
-			setData(nextData);
-		});
-	}, [listToken]);
-
+export function List({ data }) {
 	return (
 		<>
 			<p>
